detekt: moved the detekt configuration back
